Split Wizard's Waltz and Prologues game engines into their own C files. Delete scenes.c, and refactor src/scenes.h to include/engines.h.
