+++ include/engines.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "global.h"
+#include "sound.h"
+
+/*
+    General "Game Engine" header file.
+    Should probably contain all of the functions and/or headers that are referenced in the game engine C files,
+    along with any other sort of common data.
+*/
