+++ src/engines/prologues.h
@@ -1,0 +1,802 @@
+#pragma once
+
+#include "global.h"
+#include "sound.h"
+
+
+  // // //  PROLOGUE: SNEAKY SPIRITS  // // //
+
+// Assets:
+extern u32 D_089ea088;   // GFX Null   -  Prologue: Sneaky Spirits
+extern u32 D_089ea08c[]; // GFX Array  -  Prologue: Sneaky Spirits
+extern u32 D_088ad2bc[]; // Animation  -  Text (Sneaky Spirits)
+
+// Functions:
+extern void func_0804510c(void);    // [func_0804510c] GFX_INIT Func_02
+extern void func_0804511c(void);    // [func_0804511c] GFX_INIT Func_01
+extern void func_0804514c(void);    // [func_0804514c] GFX_INIT Func_00
+extern void func_0804517c(u32);     // [func_0804517c] MAIN - Init
+extern void func_080451d8(void);    // [func_080451d8] ENGINE Func_00   (STUB)
+extern void func_080451dc(void);    // [func_080451dc] MAIN - Update    (STUB)
+extern void func_080451e0(void);    // [func_080451e0] MAIN - Close     (STUB)
+extern void func_080451e4(u32);     // [func_080451e4] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: SPACEBALL  // // //
+
+
+// Assets:
+extern u32 D_089ea26c;   // GFX Null   -  Prologue: Spaceball
+extern u32 D_089ea270[]; // GFX Array  -  Prologue: Spaceball
+extern u32 D_088ad3f4[]; // Animation  -  Text (Spaceball)
+
+// Functions:
+extern void func_08045208(void);    // [func_08045208] GFX_INIT Func_02
+extern void func_08045218(void);    // [func_08045218] GFX_INIT Func_01
+extern void func_08045248(void);    // [func_08045248] GFX_INIT Func_00
+extern void func_08045278(u32);     // [func_08045278] MAIN - Init
+extern void func_080452d4(void);    // [func_080452d4] ENGINE Func_00   (STUB)
+extern void func_080452d8(void);    // [func_080452d8] MAIN - Update    (STUB)
+extern void func_080452dc(void);    // [func_080452dc] MAIN - Close     (STUB)
+extern void func_080452e0(u32);     // [func_080452e0] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: THE BON ODORI  // // //
+
+
+// Assets:
+extern u32 D_089ea3e4;   // GFX Null   -  Prologue: The Bon Odori
+extern u32 D_089ea3e8[]; // GFX Array  -  Prologue: The Bon Odori
+extern u32 D_088ad530[]; // Animation  -  Text (The Bon Odori)
+
+// Functions:
+extern void func_08045304(void);    // [func_08045304] GFX_INIT Func_02
+extern void func_08045314(void);    // [func_08045314] GFX_INIT Func_01
+extern void func_08045344(void);    // [func_08045344] GFX_INIT Func_00
+extern void func_08045374(u32);     // [func_08045374] MAIN - Init
+extern void func_080453d0(void);    // [func_080453d0] ENGINE Func_00   (STUB)
+extern void func_080453d4(void);    // [func_080453d4] MAIN - Update    (STUB)
+extern void func_080453d8(void);    // [func_080453d8] MAIN - Close     (STUB)
+extern void func_080453dc(void);    // [func_080453dc] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: BON DANCE  // // //
+
+
+// Assets:
+extern u32 D_089ea514;   // GFX Null   -  Prologue: Bon Dance
+extern u32 D_089ea518[]; // GFX Array  -  Prologue: Bon Dance
+extern u32 D_088ad898[]; // Animation  -  Text (Bon Dance)
+
+// Functions:
+extern void func_08045408(void);    // [func_08045408] GFX_INIT Func_02
+extern void func_08045418(void);    // [func_08045418] GFX_INIT Func_01
+extern void func_08045448(void);    // [func_08045448] GFX_INIT Func_00
+extern void func_08045478(u32);     // [func_08045478] MAIN - Init
+extern void func_080454d4(void);    // [func_080454d4] ENGINE Func_00   (STUB)
+extern void func_080454d8(void);    // [func_080454d8] MAIN - Update    (STUB)
+extern void func_080454dc(void);    // [func_080454dc] MAIN - Close     (STUB)
+extern void func_080454e0(void);    // [func_080454e0] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: KARATE MAN  // // //
+
+
+// Assets:
+extern u32 D_089ea650;   // GFX Null   -  Prologue: Karate Man
+extern u32 D_089ea654[]; // GFX Array  -  Prologue: Karate Man
+extern u32 D_088ada78[]; // Animation  -  Text (Karate Man)
+extern u32 D_088adaa8[]; // Animation  -  Kick (Karate Man)
+
+// Functions:
+extern void func_0804550c(void);    // [func_0804550c] GFX_INIT Func_02
+extern void func_0804551c(void);    // [func_0804551c] GFX_INIT Func_01
+extern void func_0804554c(void);    // [func_0804554c] GFX_INIT Func_00
+extern void func_0804557c(u32);     // [func_0804557c] MAIN - Init
+extern void func_080455f8(void);    // [func_080455f8] ENGINE Func_00   (STUB)
+extern void func_080455fc(void);    // [func_080455fc] MAIN - Update    (STUB)
+extern void func_08045600(void);    // [func_08045600] MAIN - Close     (STUB)
+extern void func_08045604(u32);     // [func_08045604] Event 0 - Set Animation Frame (Text)
+extern void func_08045628(void);    // [func_08045628] Event 1 - Play Animation (Kick)
+
+
+  // // //  PROLOGUE: COSMIC DANCE  // // //
+
+
+// Assets:
+extern u32 D_089ea864;   // GFX Null   -  Prologue: Cosmic Dance
+extern u32 D_089ea868[]; // GFX Array  -  Prologue: Cosmic Dance
+extern u32 D_088adb94[]; // Animation  -  Text (Cosmic Dance)
+extern u32 D_088adba4[]; // Animation  -  Dancers (Cosmic Dance)
+
+// Functions:
+extern void func_08045664(void);    // [func_08045664] GFX_INIT Func_02
+extern void func_08045674(void);    // [func_08045674] GFX_INIT Func_01
+extern void func_080456a4(void);    // [func_080456a4] GFX_INIT Func_00
+extern void func_080456d4(u32);     // [func_080456d4] MAIN - Init
+extern void func_08045730(void);    // [func_08045730] ENGINE Func_00   (STUB)
+extern void func_08045734(void);    // [func_08045734] MAIN - Update    (STUB)
+extern void func_08045738(void);    // [func_08045738] MAIN - Close     (STUB)
+extern void func_0804573c(void);    // [func_0804573c] Event 0 - Show Dancers
+
+
+  // // //  PROLOGUE: NIGHT WALK  // // //
+
+
+// Assets:
+extern u32 D_089ea988;   // GFX Null   -  Prologue: Night Walk
+extern u32 D_089ea98c[]; // GFX Array  -  Prologue: Night Walk
+extern u32 D_088ae2a8[]; // Animation  -  Text (Night Walk)
+extern u32 D_088ae048[]; // Animation  -  Play-yan (Night Walk)
+extern u32 D_088ae160[]; // Animation  -  Balloons (Night Walk)
+
+// Functions:
+extern void func_080457b4(void);    // [func_080457b4] GFX_INIT Func_02
+extern void func_080457c4(void);    // [func_080457c4] GFX_INIT Func_01
+extern void func_080457f4(void);    // [func_080457f4] GFX_INIT Func_00
+extern void func_08045824(u32);     // [func_08045824] MAIN - Init
+extern void func_080458e0(void);    // [func_080458e0] ENGINE Func_00   (STUB)
+extern void func_080458e4(void);    // [func_080458e4] MAIN - Update    (STUB)
+extern void func_080458e8(void);    // [func_080458e8] MAIN - Close     (STUB)
+extern void func_080458ec(u32);     // [func_080458ec] Event 0 - Play Animation (Play-yan); Event 1 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: SHOWTIME  // // //
+
+
+// Assets:
+extern u32 D_089eaadc;   // GFX Null   -  Prologue: Showtime
+extern u32 D_089eaae0[]; // GFX Array  -  Prologue: Showtime
+extern u32 D_088ae3e4[]; // Animation  -  Text (Showtime)
+extern u32 D_088ae3f4[]; // Animation  -  Penguin (Showtime)
+
+// Functions:
+extern void func_08045944(void);    // [func_08045944] GFX_INIT Func_02
+extern void func_08045954(void);    // [func_08045954] GFX_INIT Func_01
+extern void func_08045984(void);    // [func_08045984] GFX_INIT Func_00
+extern void func_080459b4(u32);     // [func_080459b4] MAIN - Init
+extern void func_08045a44(void);    // [func_08045a44] ENGINE Func_00   (STUB)
+extern void func_08045a48(void);    // [func_08045a48] MAIN - Update    (STUB)
+extern void func_08045a4c(void);    // [func_08045a4c] MAIN - Close     (STUB)
+extern void func_08045a50(void);    // [func_08045a50] Event 0 - Play Animation (Text), Play Animation (Penguin)
+
+
+  // // //  PROLOGUE: BOUNCY ROAD  // // //
+
+
+// Assets:
+extern u32 D_089eac18;   // GFX Null   -  Prologue: Bouncy Road
+extern u32 D_089eac1c[]; // GFX Array  -  Prologue: Bouncy Road
+extern u32 D_088af5a0[]; // Animation  -  Text (Bouncy Road)
+extern u32 D_088af510[]; // Animation  -  Bouncers (Bouncy Road)
+
+// Functions:
+extern void func_08045a90(void);    // [func_08045a90] GFX_INIT Func_02
+extern void func_08045aa0(void);    // [func_08045aa0] GFX_INIT Func_01
+extern void func_08045ad0(void);    // [func_08045ad0] GFX_INIT Func_00
+extern void func_08045b00(u32);     // [func_08045b00] MAIN - Init
+extern void func_08045b88(void);    // [func_08045b88] ENGINE Func_00   (STUB)
+extern void func_08045b8c(void);    // [func_08045b8c] MAIN - Update    (STUB)
+extern void func_08045b90(void);    // [func_08045b90] MAIN - Close     (STUB)
+extern void func_08045b94(u32);     // [func_08045b94] Event 0 - Play Animation (Bouncers); Event 1 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: RHYTHM TWEEZERS  // // //
+
+
+// Assets:
+extern u32 D_089ead54;   // GFX Null   -  Prologue: Rhythm Tweezers
+extern u32 D_089ead58[]; // GFX Array  -  Prologue: Rhythm Tweezers
+extern u32 D_088af6b4[]; // Animation  -  Text (Rhythm Tweezers)
+
+// Functions:
+extern void func_08045bec(void);    // [func_08045bec] GFX_INIT Func_02
+extern void func_08045bfc(void);    // [func_08045bfc] GFX_INIT Func_01
+extern void func_08045c2c(void);    // [func_08045c2c] GFX_INIT Func_00
+extern void func_08045c5c(u32);     // [func_08045c5c] MAIN - Init
+extern void func_08045cb8(void);    // [func_08045cb8] ENGINE Func_00   (STUB)
+extern void func_08045cbc(void);    // [func_08045cbc] MAIN - Update    (STUB)
+extern void func_08045cc0(void);    // [func_08045cc0] MAIN - Close     (STUB)
+extern void func_08045cc4(void);    // [func_08045cc4] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: FIREWORKS  // // //
+
+
+// Assets:
+extern u32 D_089eae84;   // GFX Null   -  Prologue: Fireworks
+extern u32 D_089eae88[]; // GFX Array  -  Prologue: Fireworks
+extern u32 D_088af7d4[]; // Animation  -  Text (Fireworks)
+
+// Functions:
+extern void func_08045cf0(void);    // [func_08045cf0] GFX_INIT Func_02
+extern void func_08045d00(void);    // [func_08045d00] GFX_INIT Func_01
+extern void func_08045d30(void);    // [func_08045d30] GFX_INIT Func_00
+extern void func_08045d60(u32);     // [func_08045d60] MAIN - Init
+extern void func_08045dbc(void);    // [func_08045dbc] ENGINE Func_00   (STUB)
+extern void func_08045dc0(void);    // [func_08045dc0] MAIN - Update    (STUB)
+extern void func_08045dc4(void);    // [func_08045dc4] MAIN - Close     (STUB)
+extern void func_08045dc8(u32);     // [func_08045dc8] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: THE CLAPPY TRIO  // // //
+
+
+// Assets:
+extern u32 D_089eb014;   // GFX Null   -  Prologue: The Clappy Trio
+extern u32 D_089eb018[]; // GFX Array  -  Prologue: The Clappy Trio
+extern u32 D_088afb9c[]; // Animation  -  Text (The Clappy Trio)
+
+// Functions:
+extern void func_08045dec(void);    // [func_08045dec] GFX_INIT Func_02
+extern void func_08045dfc(void);    // [func_08045dfc] GFX_INIT Func_01
+extern void func_08045e2c(void);    // [func_08045e2c] GFX_INIT Func_00
+extern void func_08045e5c(u32);     // [func_08045e5c] MAIN - Init
+extern void func_08045eb8(void);    // [func_08045eb8] ENGINE Func_00   (STUB)
+extern void func_08045ebc(void);    // [func_08045ebc] MAIN - Update    (STUB)
+extern void func_08045ec0(void);    // [func_08045ec0] MAIN - Close     (STUB)
+extern void func_08045ec4(u32);     // [func_08045ec4] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: THE SNAPPY TRIO  // // //
+
+
+// Assets:
+extern u32 D_089eb1c8;   // GFX Null   -  Prologue: The Snappy Trio
+extern u32 D_089eb1cc[]; // GFX Array  -  Prologue: The Snappy Trio
+extern u32 D_088afdc4[]; // Animation  -  Text (The Snappy Trio)
+
+// Functions:
+extern void func_08045ee8(void);    // [func_08045ee8] GFX_INIT Func_02
+extern void func_08045ef8(void);    // [func_08045ef8] GFX_INIT Func_01
+extern void func_08045f28(void);    // [func_08045f28] GFX_INIT Func_00
+extern void func_08045f58(u32);     // [func_08045f58] MAIN - Init
+extern void func_08045fb4(void);    // [func_08045fb4] ENGINE Func_00   (STUB)
+extern void func_08045fb8(void);    // [func_08045fb8] MAIN - Update    (STUB)
+extern void func_08045fbc(void);    // [func_08045fbc] MAIN - Close     (STUB)
+extern void func_08045fc0(u32);     // [func_08045fc0] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: SAMURAI SLICE  // // //
+
+
+// Assets:
+extern u32 D_089eb37c;   // GFX Null   -  Prologue: Samurai Slice
+extern u32 D_089eb380[]; // GFX Array  -  Prologue: Samurai Slice
+extern u32 D_088affe0[]; // Animation  -  Text (Samurai Slice)
+
+// Functions:
+extern void func_08045fe4(void);    // [func_08045fe4] GFX_INIT Func_02
+extern void func_08045ff4(void);    // [func_08045ff4] GFX_INIT Func_01
+extern void func_08046024(void);    // [func_08046024] GFX_INIT Func_00
+extern void func_08046054(u32);     // [func_08046054] MAIN - Init
+extern void func_080460c8(void);    // [func_080460c8] ENGINE Func_00   (STUB)
+extern void func_080460cc(void);    // [func_080460cc] MAIN - Update    (STUB)
+extern void func_080460d0(void);    // [func_080460d0] MAIN - Close     (STUB)
+extern void func_080460d4(u32);     // [func_080460d4] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: POWER CALLIGRAPHY  // // //
+
+
+// Assets:
+extern u32 D_089eb530;   // GFX Null   -  Prologue: Power Calligraphy
+extern u32 D_089eb534[]; // GFX Array  -  Prologue: Power Calligraphy
+extern u32 D_088b0690[]; // Animation  -  Text (Power Calligraphy)
+extern u32 D_088b06b0[]; // Animation  -  HAI (Power Calligraphy)
+extern const struct SequenceData s_sword_hi_seqData; // Sound Effect - HAI
+
+// Functions:
+extern void func_080460f8(void);    // [func_080460f8] GFX_INIT Func_02
+extern void func_08046108(void);    // [func_08046108] GFX_INIT Func_01
+extern void func_08046138(void);    // [func_08046138] GFX_INIT Func_00
+extern void func_08046168(u32);     // [func_08046168] MAIN - Init
+extern void func_080461c4(void);    // [func_080461c4] ENGINE Func_00   (STUB)
+extern void func_080461c8(void);    // [func_080461c8] MAIN - Update    (STUB)
+extern void func_080461cc(void);    // [func_080461cc] MAIN - Close     (STUB)
+extern void func_080461d0(u32);     // [func_080461d0] Event 0 - Set Animation Frame (Text)
+extern void func_080461f4(void);    // [func_080461f4] Event 1 - HAI
+
+
+  // // //  PROLOGUE: MARCHING ORDERS  // // //
+
+
+// Assets:
+extern u32 D_089eb66c;   // GFX Null   -  Prologue: Marching Orders
+extern u32 D_089eb670[]; // GFX Array  -  Prologue: Marching Orders
+extern u32 D_088b09b4[]; // Animation  -  Text (Marching Orders)
+
+// Functions:
+extern void func_08046234(void);    // [func_08046234] GFX_INIT Func_02
+extern void func_08046244(void);    // [func_08046244] GFX_INIT Func_01
+extern void func_08046274(void);    // [func_08046274] GFX_INIT Func_00
+extern void func_080462a4(u32);     // [func_080462a4] MAIN - Init
+extern void func_08046300(void);    // [func_08046300] ENGINE Func_00   (STUB)
+extern void func_08046304(void);    // [func_08046304] MAIN - Update    (STUB)
+extern void func_08046308(void);    // [func_08046308] MAIN - Close     (STUB)
+extern void func_0804630c(u32);     // [func_0804630c] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: MARCHING ORDERS 2  // // //
+
+
+// Assets:
+extern u32 D_089eb814;   // GFX Null   -  Prologue: Marching Orders 2
+extern u32 D_089eb818[]; // GFX Array  -  Prologue: Marching Orders 2
+extern u32 D_088b0bf0[]; // Animation  -  Text (Marching Orders 2)
+
+// Functions:
+extern void func_08046330(void);    // [func_08046330] GFX_INIT Func_02
+extern void func_08046340(void);    // [func_08046340] GFX_INIT Func_01
+extern void func_08046370(void);    // [func_08046370] GFX_INIT Func_00
+extern void func_080463a0(u32);     // [func_080463a0] MAIN - Init
+extern void func_080463fc(void);    // [func_080463fc] ENGINE Func_00   (STUB)
+extern void func_08046400(void);    // [func_08046400] MAIN - Update    (STUB)
+extern void func_08046404(void);    // [func_08046404] MAIN - Close     (STUB)
+extern void func_08046408(u32);     // [func_08046408] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: POLYRHYTHM  // // //
+
+
+// Assets:
+extern u32 D_089eb9bc;   // GFX Null   -  Prologue: Polyrhythm
+extern u32 D_089eb9c0[]; // GFX Array  -  Prologue: Polyrhythm
+extern u32 D_088b0cb4[]; // Animation  -  Text (Polyrhythm)
+
+// Functions:
+extern void func_0804642c(void);    // [func_0804642c] GFX_INIT Func_02
+extern void func_0804643c(void);    // [func_0804643c] GFX_INIT Func_01
+extern void func_0804646c(void);    // [func_0804646c] GFX_INIT Func_00
+extern void func_0804649c(u32);     // [func_0804649c] MAIN - Init
+extern void func_080464f8(void);    // [func_080464f8] ENGINE Func_00   (STUB)
+extern void func_080464fc(void);    // [func_080464fc] MAIN - Update    (STUB)
+extern void func_08046500(void);    // [func_08046500] MAIN - Close     (STUB)
+extern void func_08046504(u32);     // [func_08046504] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: QUIZ SHOW  // // //
+
+
+// Assets:
+extern u32 D_089ebb7c;   // GFX Null   -  Prologue: Quiz Show
+extern u32 D_089ebb80[]; // GFX Array  -  Prologue: Quiz Show
+extern u32 D_088b13c4[]; // Animation  -  Text (Quiz Show)
+
+// Functions:
+extern void func_08046528(void);    // [func_08046528] GFX_INIT Func_02
+extern void func_08046538(void);    // [func_08046538] GFX_INIT Func_01
+extern void func_08046568(void);    // [func_08046568] GFX_INIT Func_00
+extern void func_08046598(u32);     // [func_08046598] MAIN - Init
+extern void func_080465f8(void);    // [func_080465f8] ENGINE Func_00   (STUB)
+extern void func_080465fc(void);    // [func_080465fc] MAIN - Update    (STUB)
+extern void func_08046600(void);    // [func_08046600] MAIN - Close     (STUB)
+extern void func_08046604(u32);     // [func_08046604] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 1  // // //
+
+
+// Assets:
+extern u32 D_089ebcac;   // GFX Null   -  Prologue: Remix 1
+extern u32 D_089ebcb0[]; // GFX Array  -  Prologue: Remix 1
+extern u32 D_088b17d0[]; // Animation  -  Text (Remix 1)
+
+// Functions:
+extern void func_08046628(void);    // [func_08046628] GFX_INIT Func_02
+extern void func_08046638(void);    // [func_08046638] GFX_INIT Func_01
+extern void func_08046668(void);    // [func_08046668] GFX_INIT Func_00
+extern void func_08046698(u32);     // [func_08046698] MAIN - Init
+extern void func_080466f4(void);    // [func_080466f4] ENGINE Func_00   (STUB)
+extern void func_080466f8(void);    // [func_080466f8] MAIN - Update    (STUB)
+extern void func_080466fc(void);    // [func_080466fc] MAIN - Close     (STUB)
+extern void func_08046700(u32);     // [func_08046700] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 2  // // //
+
+
+// Assets:
+extern u32 D_089ebe78;   // GFX Null   -  Prologue: Remix 2
+extern u32 D_089ebe7c[]; // GFX Array  -  Prologue: Remix 2
+extern u32 D_088b1c9c[]; // Animation  -  Text (Remix 2)
+
+// Functions:
+extern void func_08046724(void);    // [func_08046724] GFX_INIT Func_02
+extern void func_08046734(void);    // [func_08046734] GFX_INIT Func_01
+extern void func_08046764(void);    // [func_08046764] GFX_INIT Func_00
+extern void func_08046794(u32);     // [func_08046794] MAIN - Init
+extern void func_080467f0(void);    // [func_080467f0] ENGINE Func_00   (STUB)
+extern void func_080467f4(void);    // [func_080467f4] MAIN - Update    (STUB)
+extern void func_080467f8(void);    // [func_080467f8] MAIN - Close     (STUB)
+extern void func_080467fc(u32);     // [func_080467fc] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 3  // // //
+
+
+// Assets:
+extern u32 D_089ec044;   // GFX Null   -  Prologue: Remix 3
+extern u32 D_089ec048[]; // GFX Array  -  Prologue: Remix 3
+extern u32 D_088b2044[]; // Animation  -  Text (Remix 3)
+
+// Functions:
+extern void func_08046820(void);    // [func_08046820] GFX_INIT Func_02
+extern void func_08046830(void);    // [func_08046830] GFX_INIT Func_01
+extern void func_08046860(void);    // [func_08046860] GFX_INIT Func_00
+extern void func_08046890(u32);     // [func_08046890] MAIN - Init
+extern void func_080468ec(void);    // [func_080468ec] ENGINE Func_00   (STUB)
+extern void func_080468f0(void);    // [func_080468f0] MAIN - Update    (STUB)
+extern void func_080468f4(void);    // [func_080468f4] MAIN - Close     (STUB)
+extern void func_080468f8(u32);     // [func_080468f8] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 4  // // //
+
+
+// Assets:
+extern u32 D_089ec210;   // GFX Null   -  Prologue: Remix 4
+extern u32 D_089ec214[]; // GFX Array  -  Prologue: Remix 4
+extern u32 D_088b23ec[]; // Animation  -  Text (Remix 4)
+
+// Functions:
+extern void func_0804691c(void);    // [func_0804691c] GFX_INIT Func_02
+extern void func_0804692c(void);    // [func_0804692c] GFX_INIT Func_01
+extern void func_0804695c(void);    // [func_0804695c] GFX_INIT Func_00
+extern void func_0804698c(u32);     // [func_0804698c] MAIN - Init
+extern void func_080469e8(void);    // [func_080469e8] ENGINE Func_00   (STUB)
+extern void func_080469ec(void);    // [func_080469ec] MAIN - Update    (STUB)
+extern void func_080469f0(void);    // [func_080469f0] MAIN - Close     (STUB)
+extern void func_080469f4(u32);     // [func_080469f4] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 5  // // //
+
+
+// Assets:
+extern u32 D_089ec3e8;   // GFX Null   -  Prologue: Remix 5
+extern u32 D_089ec3ec[]; // GFX Array  -  Prologue: Remix 5
+extern u32 D_088b27f4[]; // Animation  -  Text (Remix 5)
+
+// Functions:
+extern void func_08046a18(void);    // [func_08046a18] GFX_INIT Func_02
+extern void func_08046a28(void);    // [func_08046a28] GFX_INIT Func_01
+extern void func_08046a58(void);    // [func_08046a58] GFX_INIT Func_00
+extern void func_08046a88(u32);     // [func_08046a88] MAIN - Init
+extern void func_08046ae4(void);    // [func_08046ae4] ENGINE Func_00   (STUB)
+extern void func_08046ae8(void);    // [func_08046ae8] MAIN - Update    (STUB)
+extern void func_08046aec(void);    // [func_08046aec] MAIN - Close     (STUB)
+extern void func_08046af0(u32);     // [func_08046af0] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 6  // // //
+
+
+// Assets:
+extern u32 D_089ec5c0;   // GFX Null   -  Prologue: Remix 6
+extern u32 D_089ec5c4[]; // GFX Array  -  Prologue: Remix 6
+extern u32 D_088b2b9c[]; // Animation  -  Text (Remix 6)
+
+// Functions:
+extern void func_08046b14(void);    // [func_08046b14] GFX_INIT Func_02
+extern void func_08046b24(void);    // [func_08046b24] GFX_INIT Func_01
+extern void func_08046b54(void);    // [func_08046b54] GFX_INIT Func_00
+extern void func_08046b84(u32);     // [func_08046b84] MAIN - Init
+extern void func_08046be0(void);    // [func_08046be0] ENGINE Func_00   (STUB)
+extern void func_08046be4(void);    // [func_08046be4] MAIN - Update    (STUB)
+extern void func_08046be8(void);    // [func_08046be8] MAIN - Close     (STUB)
+extern void func_08046bec(u32);     // [func_08046af0] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 7  // // //
+
+
+// Assets:
+extern u32 D_089ec7b0;   // GFX Null   -  Prologue: Remix 7
+extern u32 D_089ec7b4[]; // GFX Array  -  Prologue: Remix 7
+extern u32 D_088b2f44[]; // Animation  -  Text (Remix 7)
+
+// Functions:
+extern void func_08046c10(void);    // [func_08046c10] GFX_INIT Func_02
+extern void func_08046c20(void);    // [func_08046c20] GFX_INIT Func_01
+extern void func_08046c50(void);    // [func_08046c50] GFX_INIT Func_00
+extern void func_08046c80(u32);     // [func_08046c80] MAIN - Init
+extern void func_08046cdc(void);    // [func_08046cdc] ENGINE Func_00   (STUB)
+extern void func_08046ce0(void);    // [func_08046ce0] MAIN - Update    (STUB)
+extern void func_08046ce4(void);    // [func_08046ce4] MAIN - Close     (STUB)
+extern void func_08046ce8(u32);     // [func_08046ce8] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: REMIX 8  // // //
+
+
+// Assets:
+extern u32 D_089ec988;   // GFX Null   -  Prologue: Remix 8
+extern u32 D_089ec98c[]; // GFX Array  -  Prologue: Remix 8
+extern u32 D_088b332c[]; // Animation  -  Text (Remix 8)
+
+// Functions:
+extern void func_08046d0c(void);    // [func_08046d0c] GFX_INIT Func_02
+extern void func_08046d1c(void);    // [func_08046d1c] GFX_INIT Func_01
+extern void func_08046d4c(void);    // [func_08046d4c] GFX_INIT Func_00
+extern void func_08046d7c(u32);     // [func_08046d7c] MAIN - Init
+extern void func_08046dd8(void);    // [func_08046dd8] ENGINE Func_00   (STUB)
+extern void func_08046ddc(void);    // [func_08046ddc] MAIN - Update    (STUB)
+extern void func_08046de0(void);    // [func_08046de0] MAIN - Close     (STUB)
+extern void func_08046de4(u32);     // [func_08046de4] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: BUNNY HOP  // // //
+
+
+// Assets:
+extern u32 D_089ecb60;   // GFX Null   -  Prologue: Bunny Hop
+extern u32 D_089ecb64[]; // GFX Array  -  Prologue: Bunny Hop
+extern u32 D_088b35dc[]; // Animation  -  Text (Bunny Hop)
+extern u32 D_088b35a4[]; // Animation  -  Rabbits (Bunny Hop)
+
+// Functions:
+extern void func_08046e08(void);    // [func_08046e08] GFX_INIT Func_02
+extern void func_08046e18(void);    // [func_08046e18] GFX_INIT Func_01
+extern void func_08046e48(void);    // [func_08046e48] GFX_INIT Func_00
+extern void func_08046e78(u32);     // [func_08046e78] MAIN - Init
+extern void func_08046f00(void);    // [func_08046f00] ENGINE Func_00   (STUB)
+extern void func_08046f04(void);    // [func_08046f04] MAIN - Update    (STUB)
+extern void func_08046f08(void);    // [func_08046f08] MAIN - Close     (STUB)
+extern void func_08046f0c(u32);     // [func_08046f0c] Event 0 - Set Animation Frame (Text)
+extern void func_08046f30(void);    // [func_08046f30] Event 1 - Play Animation (Rabbits)
+
+
+  // // //  PROLOGUE: RAP MEN  // // //
+
+
+// Assets:
+extern u32 D_089ecd44;   // GFX Null   -  Prologue: Rap Men
+extern u32 D_089ecd48[]; // GFX Array  -  Prologue: Rap Men
+extern u32 D_088b37d8[]; // Animation  -  Text (Rap Men)
+
+// Functions:
+extern void func_08046f5c(void);    // [func_08046f5c] GFX_INIT Func_02
+extern void func_08046f6c(void);    // [func_08046f6c] GFX_INIT Func_01
+extern void func_08046f9c(void);    // [func_08046f9c] GFX_INIT Func_00
+extern void func_08046fcc(u32);     // [func_08046fcc] MAIN - Init
+extern void func_08047028(void);    // [func_08047028] ENGINE Func_00   (STUB)
+extern void func_0804702c(void);    // [func_0804702c] MAIN - Update    (STUB)
+extern void func_08047030(void);    // [func_08047030] MAIN - Close     (STUB)
+extern void func_08047034(void);    // [func_08047034] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: RAP WOMEN  // // //
+
+
+// Assets:
+extern u32 D_089ece80;   // GFX Null   -  Prologue: Rap Women
+extern u32 D_089ece84[]; // GFX Array  -  Prologue: Rap Women
+extern u32 D_088b3ac4[]; // Animation  -  Text (Rap Women)
+
+// Functions:
+extern void func_08047060(void);    // [func_08047060] GFX_INIT Func_02
+extern void func_08047070(void);    // [func_08047070] GFX_INIT Func_01
+extern void func_080470a0(void);    // [func_080470a0] GFX_INIT Func_00
+extern void func_080470d0(u32);     // [func_080470d0] MAIN - Init
+extern void func_0804712c(void);    // [func_0804712c] ENGINE Func_00   (STUB)
+extern void func_08047130(void);    // [func_08047130] MAIN - Update    (STUB)
+extern void func_08047134(void);    // [func_08047134] MAIN - Close     (STUB)
+extern void func_08047138(void);    // [func_08047138] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: RAT RACE  // // //
+
+
+// Assets:
+extern u32 D_089ecfbc;   // GFX Null   -  Prologue: Rat Race
+extern u32 D_089ecfc0[]; // GFX Array  -  Prologue: Rat Race
+extern u32 D_088b3d54[]; // Animation  -  Text (Rat Race)
+extern u32 D_088b3d1c[]; // Animation  -  Rat (Rat Race)
+
+// Functions:
+extern void func_08047164(void);    // [func_08047164] GFX_INIT Func_02
+extern void func_08047174(void);    // [func_08047174] GFX_INIT Func_01
+extern void func_080471a4(void);    // [func_080471a4] GFX_INIT Func_00
+extern void func_080471d4(u32);     // [func_080471d4] MAIN - Init
+extern void func_08047250(void);    // [func_08047250] ENGINE Func_00   (STUB)
+extern void func_08047254(void);    // [func_08047254] MAIN - Update    (STUB)
+extern void func_08047258(void);    // [func_08047258] MAIN - Close     (STUB)
+extern void func_0804725c(u32);     // [func_0804725c] Event 0 - Set Animation Frame (Text)
+extern void func_08047280(void);    // [func_08047280] Event 1 - Play Animation (Rat)
+
+
+  // // //  PROLOGUE: NINJA BODYGUARD  // // //
+
+
+// Assets:
+extern u32 D_089ed188;   // GFX Null   -  Prologue: Ninja Bodyguard
+extern u32 D_089ed18c[]; // GFX Array  -  Prologue: Ninja Bodyguard
+extern u32 D_088b4054[]; // Animation  -  Text (Ninja Bodyguard)
+
+// Functions:
+extern void func_080472bc(void);    // [func_080472bc] GFX_INIT Func_02
+extern void func_080472cc(void);    // [func_080472cc] GFX_INIT Func_01
+extern void func_080472fc(void);    // [func_080472fc] GFX_INIT Func_00
+extern void func_0804732c(u32);     // [func_0804732c] MAIN - Init
+extern void func_08047398(void);    // [func_08047398] ENGINE Func_00   (STUB)
+extern void func_0804739c(void);    // [func_0804739c] MAIN - Update    (STUB)
+extern void func_080473a0(void);    // [func_080473a0] MAIN - Close     (STUB)
+extern void func_080473a4(u32);     // [func_080473a4] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: NINJA BODYGUARD 2  // // //
+
+
+// Assets:
+extern u32 D_089ed2c4;   // GFX Null   -  Prologue: Ninja Bodyguard 2
+extern u32 D_089ed2c8[]; // GFX Array  -  Prologue: Ninja Bodyguard 2
+extern u32 D_088b432c[]; // Animation  -  Text (Ninja Bodyguard 2)
+
+// Functions:
+extern void func_080473d4(void);    // [func_080473d4] GFX_INIT Func_02
+extern void func_080473e4(void);    // [func_080473e4] GFX_INIT Func_01
+extern void func_08047414(void);    // [func_08047414] GFX_INIT Func_00
+extern void func_08047444(u32);     // [func_08047444] MAIN - Init
+extern void func_080474a0(void);    // [func_080474a0] ENGINE Func_00   (STUB)
+extern void func_080474a4(void);    // [func_080474a4] MAIN - Update    (STUB)
+extern void func_080474a8(void);    // [func_080474a8] MAIN - Close     (STUB)
+extern void func_080474ac(u32);     // [func_080474ac] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: SPACE DANCE  // // //
+
+
+// Assets:
+extern u32 D_089ed400;   // GFX Null   -  Prologue: Space Dance
+extern u32 D_089ed404[]; // GFX Array  -  Prologue: Space Dance
+extern u32 D_088b7338[]; // Animation  -  Text (Space Dance)
+extern u32 D_088b7690[]; // Animation  -  Faces (Space Dance)
+extern u32 D_088b7660[]; // Animation  -  Star (Space Dance)
+
+// Functions:
+extern void func_080474dc(void);    // [func_080474dc] GFX_INIT Func_02
+extern void func_080474ec(void);    // [func_080474ec] GFX_INIT Func_01
+extern void func_0804751c(void);    // [func_0804751c] GFX_INIT Func_00
+extern void func_0804754c(u32);     // [func_0804754c] MAIN - Init
+extern void func_08047624(void);    // [func_08047624] ENGINE Func_00   (STUB)
+extern void func_08047628(void);    // [func_08047628] MAIN - Update    (STUB)
+extern void func_0804762c(void);    // [func_0804762c] MAIN - Close     (STUB)
+extern void func_08047630(u32);     // [func_08047630] Event 0 - Animate Entity { 0 = Text; 1 = Faces; 2 = Star }
+
+
+  // // //  PROLOGUE: TAP TRIAL  // // //
+
+
+// Assets:
+extern u32 D_089ed548;   // GFX Null   -  Prologue: Tap Trial
+extern u32 D_089ed54c[]; // GFX Array  -  Prologue: Tap Trial
+extern u32 D_088b794c[]; // Animation  -  Text (Tap Trial)
+
+// Functions:
+extern void func_08047678(void);    // [func_08047678] GFX_INIT Func_02
+extern void func_08047688(void);    // [func_08047688] GFX_INIT Func_01
+extern void func_080476b8(void);    // [func_080476b8] GFX_INIT Func_00
+extern void func_080476e8(u32);     // [func_080476e8] MAIN - Init
+extern void func_08047764(void);    // [func_08047764] ENGINE Func_00   (STUB)
+extern void func_08047768(void);    // [func_08047768] MAIN - Update    (STUB)
+extern void func_0804776c(void);    // [func_0804776c] MAIN - Close     (STUB)
+extern void func_08047770(u32);     // [func_08047770] Event 0 - Set Animation Frame (Text)
+extern void func_08047794(void);    // [func_08047794] Event 1 - Animate Text
+
+
+  // // //  PROLOGUE: TAP TRIAL 2  // // //
+
+
+// Assets:
+extern u32 D_089ed6b4;   // GFX Null   -  Prologue: Tap Trial 2
+extern u32 D_089ed6b8[]; // GFX Array  -  Prologue: Tap Trial 2
+extern u32 D_088b7c88[]; // Animation  -  Text (Tap Trial 2)
+
+// Functions:
+extern void func_080477b4(void);    // [func_080477b4] GFX_INIT Func_02
+extern void func_080477c4(void);    // [func_080477c4] GFX_INIT Func_01
+extern void func_080477f4(void);    // [func_080477f4] GFX_INIT Func_00
+extern void func_08047824(u32);     // [func_08047824] MAIN - Init
+extern void func_08047888(void);    // [func_08047888] ENGINE Func_00   (STUB)
+extern void func_0804788c(void);    // [func_0804788c] MAIN - Update    (STUB)
+extern void func_08047890(void);    // [func_08047890] MAIN - Close     (STUB)
+extern void func_08047894(u32);     // [func_08047894] Event 0 - Set Animation Frame (Text)
+extern void func_080478b8(void);    // [func_080478b8] Event 1 - Animate Text
+
+
+  // // //  PROLOGUE: TOSS BOYS  // // //
+
+
+// Assets:
+extern u32 D_089ed820;   // GFX Null   -  Prologue: Toss Boys
+extern u32 D_089ed824[]; // GFX Array  -  Prologue: Toss Boys
+extern u32 D_088b8098[]; // Animation  -  Text (Toss Boys)
+
+// Functions:
+extern void func_080478d8(void);    // [func_080478d8] GFX_INIT Func_02
+extern void func_080478e8(void);    // [func_080478e8] GFX_INIT Func_01
+extern void func_08047918(void);    // [func_08047918] GFX_INIT Func_00
+extern void func_08047948(u32);     // [func_08047948] MAIN - Init
+extern void func_080479a4(void);    // [func_080479a4] ENGINE Func_00   (STUB)
+extern void func_080479a8(void);    // [func_080479a8] MAIN - Update    (STUB)
+extern void func_080479ac(void);    // [func_080479ac] MAIN - Close     (STUB)
+extern void func_080479b0(u32);     // [func_080479b0] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: TOSS BOYS 2  // // //
+
+
+// Assets:
+extern u32 D_089ed9ec;   // GFX Null   -  Prologue: Toss Boys 2
+extern u32 D_089ed9f0[]; // GFX Array  -  Prologue: Toss Boys 2
+extern u32 D_088b8388[]; // Animation  -  Text (Toss Boys 2)
+
+// Functions:
+extern void func_080479d4(void);    // [func_080479d4] GFX_INIT Func_02
+extern void func_080479e4(void);    // [func_080479e4] GFX_INIT Func_01
+extern void func_08047a14(void);    // [func_08047a14] GFX_INIT Func_00
+extern void func_08047a44(u32);     // [func_08047a44] MAIN - Init
+extern void func_08047aa0(void);    // [func_08047aa0] ENGINE Func_00   (STUB)
+extern void func_08047aa4(void);    // [func_08047aa4] MAIN - Update    (STUB)
+extern void func_08047aa8(void);    // [func_08047aa8] MAIN - Close     (STUB)
+extern void func_08047aac(u32);     // [func_08047aac] Event 0 - Set Animation Frame (Text)
+
+
+  // // //  PROLOGUE: TRAM & PAULINE  // // //
+
+
+// Assets:
+extern u32 D_089edbb8;   // GFX Null   -  Prologue: Tram & Pauline
+extern u32 D_089edbbc[]; // GFX Array  -  Prologue: Tram & Pauline
+extern u32 D_088b8844[]; // Animation  -  Text (Tram & Pauline)
+extern u32 D_088b8834[]; // Animation  -  Tram & Pauline (Tram & Pauline)
+
+// Functions:
+extern void func_08047ad0(void);    // [func_08047ad0] GFX_INIT Func_02
+extern void func_08047ae0(void);    // [func_08047ae0] GFX_INIT Func_01
+extern void func_08047b10(void);    // [func_08047b10] GFX_INIT Func_00
+extern void func_08047b40(u32);     // [func_08047b40] MAIN - Init
+extern void func_08047bd0(void);    // [func_08047bd0] ENGINE Func_00   (STUB)
+extern void func_08047bd4(void);    // [func_08047bd4] MAIN - Update    (STUB)
+extern void func_08047bd8(void);    // [func_08047bd8] MAIN - Close     (STUB)
+extern void func_08047bdc(void);    // [func_08047bdc] Event 0 - Play Animation (Text)
+
+
+  // // //  PROLOGUE: SICK BEATS  // // //
+
+
+// Assets:
+extern u32 D_089edcf4;   // GFX Null   -  Prologue: Sick Beats
+extern u32 D_089edcf8[]; // GFX Array  -  Prologue: Sick Beats
+extern u32 D_088b8cfc[]; // Animation  -  Text (Sick Beats)
+extern u32 D_088b8cbc[]; // Animation  -  Virus (Sick Beats)
+
+// Functions:
+extern void func_08047c08(void);    // [func_08047c08] GFX_INIT Func_02
+extern void func_08047c18(void);    // [func_08047c18] GFX_INIT Func_01
+extern void func_08047c48(void);    // [func_08047c48] GFX_INIT Func_00
+extern void func_08047c78(u32);     // [func_08047c78] MAIN - Init
+extern void func_08047d00(void);    // [func_08047d00] ENGINE Func_00   (STUB)
+extern void func_08047d04(void);    // [func_08047d04] MAIN - Update    (STUB)
+extern void func_08047d08(void);    // [func_08047d08] MAIN - Close     (STUB)
+extern void func_08047d0c(void);    // [func_08047d0c] Event 0 - Play Animation (Text), Play Animation (Virus)
+
+
+  // // //  PROLOGUE: WIZARD'S WALTZ  // // //
+
+
+// Assets:
+extern u32 D_089ede24;   // GFX Null   -  Prologue: Wizard's Waltz
+extern u32 D_089ede28[]; // GFX Array  -  Prologue: Wizard's Waltz
+extern u32 D_088b8e50[]; // Animation  -  Text (Wizard's Waltz)
+
+// Functions:
+extern void func_08047d4c(void);    // [func_08047d4c] GFX_INIT Func_02
+extern void func_08047d5c(void);    // [func_08047d5c] GFX_INIT Func_01
+extern void func_08047d8c(void);    // [func_08047d8c] GFX_INIT Func_00
+extern void func_08047dbc(u32);     // [func_08047dbc] MAIN - Init
+extern void func_08047e18(void);    // [func_08047e18] ENGINE Func_00   (STUB)
+extern void func_08047e1c(void);    // [func_08047e1c] MAIN - Update    (STUB)
+extern void func_08047e20(void);    // [func_08047e20] MAIN - Close     (STUB)
+extern void func_08047e24(void);    // [func_08047e24] Event 0 - Play Animation (Text)
