--- src/scenes.c
@@ -1,14 +1,0 @@
-#include "scenes.h"
-
-#include "src/code_08001360.h"
-#include "src/code_08003980.h"
-#include "src/code_08007468.h"
-#include "src/code_0800b778.h"
-#include "src/code_08016e18.h"
-#include "src/lib_0804c870.h"
-
-asm(".include \"include/gba.inc\"");//Temporary
-
-#include "src/scenes/wizards_waltz.inc.c"
-
-#include "src/scenes/prologues.inc.c"
