--- src/scenes.h
@@ -1,483 +1,0 @@
-#pragma once
-
-#include "global.h"
-#include "sound.h"
-
-// Wizard's Waltz
-
-extern void func_080449a4(void);        // [func_080449a4] GFX_INIT Func_02
-extern void func_080449b4(void);        // [func_080449b4] GFX_INIT Func_01
-extern void func_080449e4(void);        // [func_080449e4] GFX_INIT Func_00
-extern void func_08044a10(u32);         // [func_08044a10] MAIN - Init
-extern void func_08044b80(u32);         // [func_08044b80] ENGINE Func_00 - Set Rotation Interval
-extern void func_08044ba8(struct AffineSprite *, s32, s32, u32);  // [func_08044ba8] Update Sprite Position
-extern void func_08044c04(void);        // [func_08044c04] MAIN - Update
-extern void func_08044e60(u32);         // [func_08044e60] ENGINE Func_01 - Set Tutorial Flag
-extern void func_08044e74_stub(void);   // [func_08044e74] MAIN - Close (STUB)
-extern void func_08044e78(u32, struct WizardsWaltzCue *, u32);  // [func_08044e78] CUE - Spawn
-extern u32  func_08044f94(u32, struct WizardsWaltzCue *, u32);  // [func_08044f94] CUE - Update
-extern void func_08044fc0(u32, struct WizardsWaltzCue *, u32);  // [func_08044fc0] CUE - Despawn
-extern void func_08044fcc(u32, struct WizardsWaltzCue *, u32);  // [func_08044fcc] CUE - Hit
-extern void func_0804503c(u32, struct WizardsWaltzCue *, u32);  // [func_0804503c] CUE - Barely
-extern void func_080450d0(u32, struct WizardsWaltzCue *, u32);  // [func_080450d0] CUE - Miss
-extern void func_080450dc_stub(void);   // [func_080450dc] MAIN - Input Event (STUB)
-extern void func_080450e0_stub(void);   // [func_080450e0] COMMON Func_00 - STUB
-extern void func_080450e4_stub(void);   // [func_080450e4] COMMON Func_01 - STUB
-extern void func_080450e8(u32);         // [func_080450e8] COMMON Func_02 - Unknown
-
-// Prologue: Sneaky Spirits
-
-extern void func_0804510c(void);    // [func_0804510c] GFX_INIT Func_02
-extern void func_0804511c(void);    // [func_0804511c] GFX_INIT Func_01
-extern void func_0804514c(void);    // [func_0804514c] GFX_INIT Func_00
-extern void func_0804517c(u32);     // [func_0804517c] MAIN - Init
-extern void func_080451d8(void);    // [func_080451d8] ENGINE Func_00   (STUB)
-extern void func_080451dc(void);    // [func_080451dc] MAIN - Update    (STUB)
-extern void func_080451e0(void);    // [func_080451e0] MAIN - Close     (STUB)
-extern void func_080451e4(u32);     // [func_080451e4] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Spaceball
-
-extern void func_08045208(void);    // [func_08045208] GFX_INIT Func_02
-extern void func_08045218(void);    // [func_08045218] GFX_INIT Func_01
-extern void func_08045248(void);    // [func_08045248] GFX_INIT Func_00
-extern void func_08045278(u32);     // [func_08045278] MAIN - Init
-extern void func_080452d4(void);    // [func_080452d4] ENGINE Func_00   (STUB)
-extern void func_080452d8(void);    // [func_080452d8] MAIN - Update    (STUB)
-extern void func_080452dc(void);    // [func_080452dc] MAIN - Close     (STUB)
-extern void func_080452e0(u32);     // [func_080452e0] Event 0 - Set Animation Frame (Text)
-
-// Prologue: The Bon Odori
-
-extern void func_08045304(void);    // [func_08045304] GFX_INIT Func_02
-extern void func_08045314(void);    // [func_08045314] GFX_INIT Func_01
-extern void func_08045344(void);    // [func_08045344] GFX_INIT Func_00
-extern void func_08045374(u32);     // [func_08045374] MAIN - Init
-extern void func_080453d0(void);    // [func_080453d0] ENGINE Func_00   (STUB)
-extern void func_080453d4(void);    // [func_080453d4] MAIN - Update    (STUB)
-extern void func_080453d8(void);    // [func_080453d8] MAIN - Close     (STUB)
-extern void func_080453dc(void);    // [func_080453dc] Event 0 - Play Animation (Text)
-
-// Prologue: Bon Dance
-
-extern void func_08045408(void);    // [func_08045408] GFX_INIT Func_02
-extern void func_08045418(void);    // [func_08045418] GFX_INIT Func_01
-extern void func_08045448(void);    // [func_08045448] GFX_INIT Func_00
-extern void func_08045478(u32);     // [func_08045478] MAIN - Init
-extern void func_080454d4(void);    // [func_080454d4] ENGINE Func_00   (STUB)
-extern void func_080454d8(void);    // [func_080454d8] MAIN - Update    (STUB)
-extern void func_080454dc(void);    // [func_080454dc] MAIN - Close     (STUB)
-extern void func_080454e0(void);    // [func_080454e0] Event 0 - Play Animation (Text)
-
-// Prologue: Karate Man
-
-extern void func_0804550c(void);    // [func_0804550c] GFX_INIT Func_02
-extern void func_0804551c(void);    // [func_0804551c] GFX_INIT Func_01
-extern void func_0804554c(void);    // [func_0804554c] GFX_INIT Func_00
-extern void func_0804557c(u32);     // [func_0804557c] MAIN - Init
-extern void func_080455f8(void);    // [func_080455f8] ENGINE Func_00   (STUB)
-extern void func_080455fc(void);    // [func_080455fc] MAIN - Update    (STUB)
-extern void func_08045600(void);    // [func_08045600] MAIN - Close     (STUB)
-extern void func_08045604(u32);     // [func_08045604] Event 0 - Set Animation Frame (Text)
-extern void func_08045628(void);    // [func_08045628] Event 1 - Play Animation (Kick)
-
-// Prologue: Cosmic Dance
-
-extern void func_08045664(void);    // [func_08045664] GFX_INIT Func_02
-extern void func_08045674(void);    // [func_08045674] GFX_INIT Func_01
-extern void func_080456a4(void);    // [func_080456a4] GFX_INIT Func_00
-extern void func_080456d4(u32);     // [func_080456d4] MAIN - Init
-extern void func_08045730(void);    // [func_08045730] ENGINE Func_00   (STUB)
-extern void func_08045734(void);    // [func_08045734] MAIN - Update    (STUB)
-extern void func_08045738(void);    // [func_08045738] MAIN - Close     (STUB)
-extern void func_0804573c(void);    // [func_0804573c] Event 0 - Show Dancers
-
-// Prologue: Night Walk
-
-extern void func_080457b4(void);    // [func_080457b4] GFX_INIT Func_02
-extern void func_080457c4(void);    // [func_080457c4] GFX_INIT Func_01
-extern void func_080457f4(void);    // [func_080457f4] GFX_INIT Func_00
-extern void func_08045824(u32);     // [func_08045824] MAIN - Init
-extern void func_080458e0(void);    // [func_080458e0] ENGINE Func_00   (STUB)
-extern void func_080458e4(void);    // [func_080458e4] MAIN - Update    (STUB)
-extern void func_080458e8(void);    // [func_080458e8] MAIN - Close     (STUB)
-extern void func_080458ec(u32);     // [func_080458ec] Event 0 - Play Animation (Play-yan); Event 1 - Play Animation (Text)
-
-// Prologue: Showtime
-
-extern void func_08045944(void);    // [func_08045944] GFX_INIT Func_02
-extern void func_08045954(void);    // [func_08045954] GFX_INIT Func_01
-extern void func_08045984(void);    // [func_08045984] GFX_INIT Func_00
-extern void func_080459b4(u32);     // [func_080459b4] MAIN - Init
-extern void func_08045a44(void);    // [func_08045a44] ENGINE Func_00   (STUB)
-extern void func_08045a48(void);    // [func_08045a48] MAIN - Update    (STUB)
-extern void func_08045a4c(void);    // [func_08045a4c] MAIN - Close     (STUB)
-extern void func_08045a50(void);    // [func_08045a50] Event 0 - Play Animation (Text), Play Animation (Penguin)
-
-// Prologue: Bouncy Road
-
-extern void func_08045a90(void);    // [func_08045a90] GFX_INIT Func_02
-extern void func_08045aa0(void);    // [func_08045aa0] GFX_INIT Func_01
-extern void func_08045ad0(void);    // [func_08045ad0] GFX_INIT Func_00
-extern void func_08045b00(u32);     // [func_08045b00] MAIN - Init
-extern void func_08045b88(void);    // [func_08045b88] ENGINE Func_00   (STUB)
-extern void func_08045b8c(void);    // [func_08045b8c] MAIN - Update    (STUB)
-extern void func_08045b90(void);    // [func_08045b90] MAIN - Close     (STUB)
-extern void func_08045b94(u32);     // [func_08045b94] Event 0 - Play Animation (Bouncers); Event 1 - Play Animation (Text)
-
-// Prologue: Rhythm Tweezers
-
-extern void func_08045bec(void);    // [func_08045bec] GFX_INIT Func_02
-extern void func_08045bfc(void);    // [func_08045bfc] GFX_INIT Func_01
-extern void func_08045c2c(void);    // [func_08045c2c] GFX_INIT Func_00
-extern void func_08045c5c(u32);     // [func_08045c5c] MAIN - Init
-extern void func_08045cb8(void);    // [func_08045cb8] ENGINE Func_00   (STUB)
-extern void func_08045cbc(void);    // [func_08045cbc] MAIN - Update    (STUB)
-extern void func_08045cc0(void);    // [func_08045cc0] MAIN - Close     (STUB)
-extern void func_08045cc4(void);    // [func_08045cc4] Event 0 - Play Animation (Text)
-
-// Prologue: Fireworks
-
-extern void func_08045cf0(void);    // [func_08045cf0] GFX_INIT Func_02
-extern void func_08045d00(void);    // [func_08045d00] GFX_INIT Func_01
-extern void func_08045d30(void);    // [func_08045d30] GFX_INIT Func_00
-extern void func_08045d60(u32);     // [func_08045d60] MAIN - Init
-extern void func_08045dbc(void);    // [func_08045dbc] ENGINE Func_00   (STUB)
-extern void func_08045dc0(void);    // [func_08045dc0] MAIN - Update    (STUB)
-extern void func_08045dc4(void);    // [func_08045dc4] MAIN - Close     (STUB)
-extern void func_08045dc8(u32);     // [func_08045dc8] Event 0 - Set Animation Frame (Text)
-
-// Prologue: The Clappy Trio
-
-extern void func_08045dec(void);    // [func_08045dec] GFX_INIT Func_02
-extern void func_08045dfc(void);    // [func_08045dfc] GFX_INIT Func_01
-extern void func_08045e2c(void);    // [func_08045e2c] GFX_INIT Func_00
-extern void func_08045e5c(u32);     // [func_08045e5c] MAIN - Init
-extern void func_08045eb8(void);    // [func_08045eb8] ENGINE Func_00   (STUB)
-extern void func_08045ebc(void);    // [func_08045ebc] MAIN - Update    (STUB)
-extern void func_08045ec0(void);    // [func_08045ec0] MAIN - Close     (STUB)
-extern void func_08045ec4(u32);     // [func_08045ec4] Event 0 - Set Animation Frame (Text)
-
-// Prologue: The Snappy Trio
-
-extern void func_08045ee8(void);    // [func_08045ee8] GFX_INIT Func_02
-extern void func_08045ef8(void);    // [func_08045ef8] GFX_INIT Func_01
-extern void func_08045f28(void);    // [func_08045f28] GFX_INIT Func_00
-extern void func_08045f58(u32);     // [func_08045f58] MAIN - Init
-extern void func_08045fb4(void);    // [func_08045fb4] ENGINE Func_00   (STUB)
-extern void func_08045fb8(void);    // [func_08045fb8] MAIN - Update    (STUB)
-extern void func_08045fbc(void);    // [func_08045fbc] MAIN - Close     (STUB)
-extern void func_08045fc0(u32);     // [func_08045fc0] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Samurai Slice
-
-extern void func_08045fe4(void);    // [func_08045fe4] GFX_INIT Func_02
-extern void func_08045ff4(void);    // [func_08045ff4] GFX_INIT Func_01
-extern void func_08046024(void);    // [func_08046024] GFX_INIT Func_00
-extern void func_08046054(u32);     // [func_08046054] MAIN - Init
-extern void func_080460c8(void);    // [func_080460c8] ENGINE Func_00   (STUB)
-extern void func_080460cc(void);    // [func_080460cc] MAIN - Update    (STUB)
-extern void func_080460d0(void);    // [func_080460d0] MAIN - Close     (STUB)
-extern void func_080460d4(u32);     // [func_080460d4] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Power Calligraphy
-
-extern void func_080460f8(void);    // [func_080460f8] GFX_INIT Func_02
-extern void func_08046108(void);    // [func_08046108] GFX_INIT Func_01
-extern void func_08046138(void);    // [func_08046138] GFX_INIT Func_00
-extern void func_08046168(u32);     // [func_08046168] MAIN - Init
-extern void func_080461c4(void);    // [func_080461c4] ENGINE Func_00   (STUB)
-extern void func_080461c8(void);    // [func_080461c8] MAIN - Update    (STUB)
-extern void func_080461cc(void);    // [func_080461cc] MAIN - Close     (STUB)
-extern void func_080461d0(u32);     // [func_080461d0] Event 0 - Set Animation Frame (Text)
-extern void func_080461f4(void);    // [func_080461f4] Event 1 - HAI
-
-// Prologue: Marching Orders
-
-extern void func_08046234(void);    // [func_08046234] GFX_INIT Func_02
-extern void func_08046244(void);    // [func_08046244] GFX_INIT Func_01
-extern void func_08046274(void);    // [func_08046274] GFX_INIT Func_00
-extern void func_080462a4(u32);     // [func_080462a4] MAIN - Init
-extern void func_08046300(void);    // [func_08046300] ENGINE Func_00   (STUB)
-extern void func_08046304(void);    // [func_08046304] MAIN - Update    (STUB)
-extern void func_08046308(void);    // [func_08046308] MAIN - Close     (STUB)
-extern void func_0804630c(u32);     // [func_0804630c] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Marching Orders 2
-
-extern void func_08046330(void);    // [func_08046330] GFX_INIT Func_02
-extern void func_08046340(void);    // [func_08046340] GFX_INIT Func_01
-extern void func_08046370(void);    // [func_08046370] GFX_INIT Func_00
-extern void func_080463a0(u32);     // [func_080463a0] MAIN - Init
-extern void func_080463fc(void);    // [func_080463fc] ENGINE Func_00   (STUB)
-extern void func_08046400(void);    // [func_08046400] MAIN - Update    (STUB)
-extern void func_08046404(void);    // [func_08046404] MAIN - Close     (STUB)
-extern void func_08046408(u32);     // [func_08046408] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Polyrhythm
-
-extern void func_0804642c(void);    // [func_0804642c] GFX_INIT Func_02
-extern void func_0804643c(void);    // [func_0804643c] GFX_INIT Func_01
-extern void func_0804646c(void);    // [func_0804646c] GFX_INIT Func_00
-extern void func_0804649c(u32);     // [func_0804649c] MAIN - Init
-extern void func_080464f8(void);    // [func_080464f8] ENGINE Func_00   (STUB)
-extern void func_080464fc(void);    // [func_080464fc] MAIN - Update    (STUB)
-extern void func_08046500(void);    // [func_08046500] MAIN - Close     (STUB)
-extern void func_08046504(u32);     // [func_08046504] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Quiz Show
-
-extern void func_08046528(void);    // [func_08046528] GFX_INIT Func_02
-extern void func_08046538(void);    // [func_08046538] GFX_INIT Func_01
-extern void func_08046568(void);    // [func_08046568] GFX_INIT Func_00
-extern void func_08046598(u32);     // [func_08046598] MAIN - Init
-extern void func_080465f8(void);    // [func_080465f8] ENGINE Func_00   (STUB)
-extern void func_080465fc(void);    // [func_080465fc] MAIN - Update    (STUB)
-extern void func_08046600(void);    // [func_08046600] MAIN - Close     (STUB)
-extern void func_08046604(u32);     // [func_08046604] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 1
-
-extern void func_08046628(void);    // [func_08046628] GFX_INIT Func_02
-extern void func_08046638(void);    // [func_08046638] GFX_INIT Func_01
-extern void func_08046668(void);    // [func_08046668] GFX_INIT Func_00
-extern void func_08046698(u32);     // [func_08046698] MAIN - Init
-extern void func_080466f4(void);    // [func_080466f4] ENGINE Func_00   (STUB)
-extern void func_080466f8(void);    // [func_080466f8] MAIN - Update    (STUB)
-extern void func_080466fc(void);    // [func_080466fc] MAIN - Close     (STUB)
-extern void func_08046700(u32);     // [func_08046700] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 2
-
-extern void func_08046724(void);    // [func_08046724] GFX_INIT Func_02
-extern void func_08046734(void);    // [func_08046734] GFX_INIT Func_01
-extern void func_08046764(void);    // [func_08046764] GFX_INIT Func_00
-extern void func_08046794(u32);     // [func_08046794] MAIN - Init
-extern void func_080467f0(void);    // [func_080467f0] ENGINE Func_00   (STUB)
-extern void func_080467f4(void);    // [func_080467f4] MAIN - Update    (STUB)
-extern void func_080467f8(void);    // [func_080467f8] MAIN - Close     (STUB)
-extern void func_080467fc(u32);     // [func_080467fc] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 3
-
-extern void func_08046820(void);    // [func_08046820] GFX_INIT Func_02
-extern void func_08046830(void);    // [func_08046830] GFX_INIT Func_01
-extern void func_08046860(void);    // [func_08046860] GFX_INIT Func_00
-extern void func_08046890(u32);     // [func_08046890] MAIN - Init
-extern void func_080468ec(void);    // [func_080468ec] ENGINE Func_00   (STUB)
-extern void func_080468f0(void);    // [func_080468f0] MAIN - Update    (STUB)
-extern void func_080468f4(void);    // [func_080468f4] MAIN - Close     (STUB)
-extern void func_080468f8(u32);     // [func_080468f8] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 4
-
-extern void func_0804691c(void);    // [func_0804691c] GFX_INIT Func_02
-extern void func_0804692c(void);    // [func_0804692c] GFX_INIT Func_01
-extern void func_0804695c(void);    // [func_0804695c] GFX_INIT Func_00
-extern void func_0804698c(u32);     // [func_0804698c] MAIN - Init
-extern void func_080469e8(void);    // [func_080469e8] ENGINE Func_00   (STUB)
-extern void func_080469ec(void);    // [func_080469ec] MAIN - Update    (STUB)
-extern void func_080469f0(void);    // [func_080469f0] MAIN - Close     (STUB)
-extern void func_080469f4(u32);     // [func_080469f4] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 5
-
-extern void func_08046a18(void);    // [func_08046a18] GFX_INIT Func_02
-extern void func_08046a28(void);    // [func_08046a28] GFX_INIT Func_01
-extern void func_08046a58(void);    // [func_08046a58] GFX_INIT Func_00
-extern void func_08046a88(u32);     // [func_08046a88] MAIN - Init
-extern void func_08046ae4(void);    // [func_08046ae4] ENGINE Func_00   (STUB)
-extern void func_08046ae8(void);    // [func_08046ae8] MAIN - Update    (STUB)
-extern void func_08046aec(void);    // [func_08046aec] MAIN - Close     (STUB)
-extern void func_08046af0(u32);     // [func_08046af0] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 6
-
-extern void func_08046b14(void);    // [func_08046b14] GFX_INIT Func_02
-extern void func_08046b24(void);    // [func_08046b24] GFX_INIT Func_01
-extern void func_08046b54(void);    // [func_08046b54] GFX_INIT Func_00
-extern void func_08046b84(u32);     // [func_08046b84] MAIN - Init
-extern void func_08046be0(void);    // [func_08046be0] ENGINE Func_00   (STUB)
-extern void func_08046be4(void);    // [func_08046be4] MAIN - Update    (STUB)
-extern void func_08046be8(void);    // [func_08046be8] MAIN - Close     (STUB)
-extern void func_08046bec(u32);     // [func_08046af0] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 7
-
-extern void func_08046c10(void);    // [func_08046c10] GFX_INIT Func_02
-extern void func_08046c20(void);    // [func_08046c20] GFX_INIT Func_01
-extern void func_08046c50(void);    // [func_08046c50] GFX_INIT Func_00
-extern void func_08046c80(u32);     // [func_08046c80] MAIN - Init
-extern void func_08046cdc(void);    // [func_08046cdc] ENGINE Func_00   (STUB)
-extern void func_08046ce0(void);    // [func_08046ce0] MAIN - Update    (STUB)
-extern void func_08046ce4(void);    // [func_08046ce4] MAIN - Close     (STUB)
-extern void func_08046ce8(u32);     // [func_08046ce8] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Remix 8
-
-extern void func_08046d0c(void);    // [func_08046d0c] GFX_INIT Func_02
-extern void func_08046d1c(void);    // [func_08046d1c] GFX_INIT Func_01
-extern void func_08046d4c(void);    // [func_08046d4c] GFX_INIT Func_00
-extern void func_08046d7c(u32);     // [func_08046d7c] MAIN - Init
-extern void func_08046dd8(void);    // [func_08046dd8] ENGINE Func_00   (STUB)
-extern void func_08046ddc(void);    // [func_08046ddc] MAIN - Update    (STUB)
-extern void func_08046de0(void);    // [func_08046de0] MAIN - Close     (STUB)
-extern void func_08046de4(u32);     // [func_08046de4] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Bunny Hop
-
-extern void func_08046e08(void);    // [func_08046e08] GFX_INIT Func_02
-extern void func_08046e18(void);    // [func_08046e18] GFX_INIT Func_01
-extern void func_08046e48(void);    // [func_08046e48] GFX_INIT Func_00
-extern void func_08046e78(u32);     // [func_08046e78] MAIN - Init
-extern void func_08046f00(void);    // [func_08046f00] ENGINE Func_00   (STUB)
-extern void func_08046f04(void);    // [func_08046f04] MAIN - Update    (STUB)
-extern void func_08046f08(void);    // [func_08046f08] MAIN - Close     (STUB)
-extern void func_08046f0c(u32);     // [func_08046f0c] Event 0 - Set Animation Frame (Text)
-extern void func_08046f30(void);    // [func_08046f30] Event 1 - Play Animation (Rabbits)
-
-// Prologue: Rap Men
-
-extern void func_08046f5c(void);    // [func_08046f5c] GFX_INIT Func_02
-extern void func_08046f6c(void);    // [func_08046f6c] GFX_INIT Func_01
-extern void func_08046f9c(void);    // [func_08046f9c] GFX_INIT Func_00
-extern void func_08046fcc(u32);     // [func_08046fcc] MAIN - Init
-extern void func_08047028(void);    // [func_08047028] ENGINE Func_00   (STUB)
-extern void func_0804702c(void);    // [func_0804702c] MAIN - Update    (STUB)
-extern void func_08047030(void);    // [func_08047030] MAIN - Close     (STUB)
-extern void func_08047034(void);    // [func_08047034] Event 0 - Play Animation (Text)
-
-// Prologue: Rap Women
-
-extern void func_08047060(void);    // [func_08047060] GFX_INIT Func_02
-extern void func_08047070(void);    // [func_08047070] GFX_INIT Func_01
-extern void func_080470a0(void);    // [func_080470a0] GFX_INIT Func_00
-extern void func_080470d0(u32);     // [func_080470d0] MAIN - Init
-extern void func_0804712c(void);    // [func_0804712c] ENGINE Func_00   (STUB)
-extern void func_08047130(void);    // [func_08047130] MAIN - Update    (STUB)
-extern void func_08047134(void);    // [func_08047134] MAIN - Close     (STUB)
-extern void func_08047138(void);    // [func_08047138] Event 0 - Play Animation (Text)
-
-// Prologue: Rat Race
-
-extern void func_08047164(void);    // [func_08047164] GFX_INIT Func_02
-extern void func_08047174(void);    // [func_08047174] GFX_INIT Func_01
-extern void func_080471a4(void);    // [func_080471a4] GFX_INIT Func_00
-extern void func_080471d4(u32);     // [func_080471d4] MAIN - Init
-extern void func_08047250(void);    // [func_08047250] ENGINE Func_00   (STUB)
-extern void func_08047254(void);    // [func_08047254] MAIN - Update    (STUB)
-extern void func_08047258(void);    // [func_08047258] MAIN - Close     (STUB)
-extern void func_0804725c(u32);     // [func_0804725c] Event 0 - Set Animation Frame (Text)
-extern void func_08047280(void);    // [func_08047280] Event 1 - Play Animation (Rat)
-
-// Prologue: Ninja Bodyguard
-
-extern void func_080472bc(void);    // [func_080472bc] GFX_INIT Func_02
-extern void func_080472cc(void);    // [func_080472cc] GFX_INIT Func_01
-extern void func_080472fc(void);    // [func_080472fc] GFX_INIT Func_00
-extern void func_0804732c(u32);     // [func_0804732c] MAIN - Init
-extern void func_08047398(void);    // [func_08047398] ENGINE Func_00   (STUB)
-extern void func_0804739c(void);    // [func_0804739c] MAIN - Update    (STUB)
-extern void func_080473a0(void);    // [func_080473a0] MAIN - Close     (STUB)
-extern void func_080473a4(u32);     // [func_080473a4] Event 0 - Play Animation (Text)
-
-// Prologue: Ninja Bodyguard 2
-
-extern void func_080473d4(void);    // [func_080473d4] GFX_INIT Func_02
-extern void func_080473e4(void);    // [func_080473e4] GFX_INIT Func_01
-extern void func_08047414(void);    // [func_08047414] GFX_INIT Func_00
-extern void func_08047444(u32);     // [func_08047444] MAIN - Init
-extern void func_080474a0(void);    // [func_080474a0] ENGINE Func_00   (STUB)
-extern void func_080474a4(void);    // [func_080474a4] MAIN - Update    (STUB)
-extern void func_080474a8(void);    // [func_080474a8] MAIN - Close     (STUB)
-extern void func_080474ac(u32);     // [func_080474ac] Event 0 - Play Animation (Text)
-
-// Prologue: Space Dance
-
-extern void func_080474dc(void);    // [func_080474dc] GFX_INIT Func_02
-extern void func_080474ec(void);    // [func_080474ec] GFX_INIT Func_01
-extern void func_0804751c(void);    // [func_0804751c] GFX_INIT Func_00
-extern void func_0804754c(u32);     // [func_0804754c] MAIN - Init
-extern void func_08047624(void);    // [func_08047624] ENGINE Func_00   (STUB)
-extern void func_08047628(void);    // [func_08047628] MAIN - Update    (STUB)
-extern void func_0804762c(void);    // [func_0804762c] MAIN - Close     (STUB)
-extern void func_08047630(u32);     // [func_08047630] Event 0 - Animate Entity { 0 = Text; 1 = Faces; 2 = Star }
-
-// Prologue: Tap Trial
-
-extern void func_08047678(void);    // [func_08047678] GFX_INIT Func_02
-extern void func_08047688(void);    // [func_08047688] GFX_INIT Func_01
-extern void func_080476b8(void);    // [func_080476b8] GFX_INIT Func_00
-extern void func_080476e8(u32);     // [func_080476e8] MAIN - Init
-extern void func_08047764(void);    // [func_08047764] ENGINE Func_00   (STUB)
-extern void func_08047768(void);    // [func_08047768] MAIN - Update    (STUB)
-extern void func_0804776c(void);    // [func_0804776c] MAIN - Close     (STUB)
-extern void func_08047770(u32);     // [func_08047770] Event 0 - Set Animation Frame (Text)
-extern void func_08047794(void);    // [func_08047794] Event 1 - Animate Text
-
-// Prologue: Tap Trial 2
-
-extern void func_080477b4(void);    // [func_080477b4] GFX_INIT Func_02
-extern void func_080477c4(void);    // [func_080477c4] GFX_INIT Func_01
-extern void func_080477f4(void);    // [func_080477f4] GFX_INIT Func_00
-extern void func_08047824(u32);     // [func_08047824] MAIN - Init
-extern void func_08047888(void);    // [func_08047888] ENGINE Func_00   (STUB)
-extern void func_0804788c(void);    // [func_0804788c] MAIN - Update    (STUB)
-extern void func_08047890(void);    // [func_08047890] MAIN - Close     (STUB)
-extern void func_08047894(u32);     // [func_08047894] Event 0 - Set Animation Frame (Text)
-extern void func_080478b8(void);    // [func_080478b8] Event 1 - Animate Text
-
-// Prologue: Toss Boys
-
-extern void func_080478d8(void);    // [func_080478d8] GFX_INIT Func_02
-extern void func_080478e8(void);    // [func_080478e8] GFX_INIT Func_01
-extern void func_08047918(void);    // [func_08047918] GFX_INIT Func_00
-extern void func_08047948(u32);     // [func_08047948] MAIN - Init
-extern void func_080479a4(void);    // [func_080479a4] ENGINE Func_00   (STUB)
-extern void func_080479a8(void);    // [func_080479a8] MAIN - Update    (STUB)
-extern void func_080479ac(void);    // [func_080479ac] MAIN - Close     (STUB)
-extern void func_080479b0(u32);     // [func_080479b0] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Toss Boys 2
-
-extern void func_080479d4(void);    // [func_080479d4] GFX_INIT Func_02
-extern void func_080479e4(void);    // [func_080479e4] GFX_INIT Func_01
-extern void func_08047a14(void);    // [func_08047a14] GFX_INIT Func_00
-extern void func_08047a44(u32);     // [func_08047a44] MAIN - Init
-extern void func_08047aa0(void);    // [func_08047aa0] ENGINE Func_00   (STUB)
-extern void func_08047aa4(void);    // [func_08047aa4] MAIN - Update    (STUB)
-extern void func_08047aa8(void);    // [func_08047aa8] MAIN - Close     (STUB)
-extern void func_08047aac(u32);     // [func_08047aac] Event 0 - Set Animation Frame (Text)
-
-// Prologue: Tram & Pauline
-
-extern void func_08047ad0(void);    // [func_08047ad0] GFX_INIT Func_02
-extern void func_08047ae0(void);    // [func_08047ae0] GFX_INIT Func_01
-extern void func_08047b10(void);    // [func_08047b10] GFX_INIT Func_00
-extern void func_08047b40(u32);     // [func_08047b40] MAIN - Init
-extern void func_08047bd0(void);    // [func_08047bd0] ENGINE Func_00   (STUB)
-extern void func_08047bd4(void);    // [func_08047bd4] MAIN - Update    (STUB)
-extern void func_08047bd8(void);    // [func_08047bd8] MAIN - Close     (STUB)
-extern void func_08047bdc(void);    // [func_08047bdc] Event 0 - Play Animation (Text)
-
-// Prologue: Sick Beats
-
-extern void func_08047c08(void);    // [func_08047c08] GFX_INIT Func_02
-extern void func_08047c18(void);    // [func_08047c18] GFX_INIT Func_01
-extern void func_08047c48(void);    // [func_08047c48] GFX_INIT Func_00
-extern void func_08047c78(u32);     // [func_08047c78] MAIN - Init
-extern void func_08047d00(void);    // [func_08047d00] ENGINE Func_00   (STUB)
-extern void func_08047d04(void);    // [func_08047d04] MAIN - Update    (STUB)
-extern void func_08047d08(void);    // [func_08047d08] MAIN - Close     (STUB)
-extern void func_08047d0c(void);    // [func_08047d0c] Event 0 - Play Animation (Text), Play Animation (Virus)
-
-// Prologue: Wizard's Waltz
-
-extern void func_08047d4c(void);    // [func_08047d4c] GFX_INIT Func_02
-extern void func_08047d5c(void);    // [func_08047d5c] GFX_INIT Func_01
-extern void func_08047d8c(void);    // [func_08047d8c] GFX_INIT Func_00
-extern void func_08047dbc(u32);     // [func_08047dbc] MAIN - Init
-extern void func_08047e18(void);    // [func_08047e18] ENGINE Func_00   (STUB)
-extern void func_08047e1c(void);    // [func_08047e1c] MAIN - Update    (STUB)
-extern void func_08047e20(void);    // [func_08047e20] MAIN - Close     (STUB)
-extern void func_08047e24(void);    // [func_08047e24] Event 0 - Play Animation (Text)
